first mini review changes
