+++ src/main/java/net/dries007/tfc/common/entities/aquatic/TFCCod.java
@@ -35,7 +35,7 @@ public class TFCCod extends Cod implements AquaticMob
     {
         super.registerGoals();
         goalSelector.addGoal(1, new GetHookedGoal(this));
-        EntityHelpers.insertTFCAvoidGoal(this, goalSelector, 2);
+        EntityHelpers.replaceAvoidEntityGoal(this, goalSelector, 2);
     }
 
     @Override
