+++ src/main/java/net/dries007/tfc/common/entities/aquatic/TFCSalmon.java
@@ -35,7 +35,7 @@ public class TFCSalmon extends Salmon implements AquaticMob
     {
         super.registerGoals();
         goalSelector.addGoal(1, new GetHookedGoal(this));
-        EntityHelpers.insertTFCAvoidGoal(this, goalSelector, 2);
+        EntityHelpers.replaceAvoidEntityGoal(this, goalSelector, 2);
     }
 
     @Override
