+++ src/main/resources/data/tfc/tags/items/pig_food.json
@@ -2,111 +2,6 @@
   "__comment__": "This file was automatically created by mcresources",
   "replace": false,
   "values": [
-    "tfc:food/banana",
-    "tfc:food/blackberry",
-    "tfc:food/blueberry",
-    "tfc:food/bunchberry",
-    "tfc:food/cherry",
-    "tfc:food/cloudberry",
-    "tfc:food/cranberry",
-    "tfc:food/elderberry",
-    "tfc:food/gooseberry",
-    "tfc:food/green_apple",
-    "tfc:food/lemon",
-    "tfc:food/olive",
-    "tfc:food/orange",
-    "tfc:food/peach",
-    "tfc:food/plum",
-    "tfc:food/raspberry",
-    "tfc:food/red_apple",
-    "tfc:food/snowberry",
-    "tfc:food/strawberry",
-    "tfc:food/wintergreen_berry",
-    "tfc:food/barley",
-    "tfc:food/barley_grain",
-    "tfc:food/barley_flour",
-    "tfc:food/barley_dough",
-    "tfc:food/barley_bread",
-    "tfc:food/maize",
-    "tfc:food/maize_grain",
-    "tfc:food/maize_flour",
-    "tfc:food/maize_dough",
-    "tfc:food/maize_bread",
-    "tfc:food/oat",
-    "tfc:food/oat_grain",
-    "tfc:food/oat_flour",
-    "tfc:food/oat_dough",
-    "tfc:food/oat_bread",
-    "tfc:food/rice",
-    "tfc:food/rice_grain",
-    "tfc:food/rice_flour",
-    "tfc:food/rice_dough",
-    "tfc:food/rice_bread",
-    "tfc:food/rye",
-    "tfc:food/rye_grain",
-    "tfc:food/rye_flour",
-    "tfc:food/rye_dough",
-    "tfc:food/rye_bread",
-    "tfc:food/wheat",
-    "tfc:food/wheat_grain",
-    "tfc:food/wheat_flour",
-    "tfc:food/wheat_dough",
-    "tfc:food/wheat_bread",
-    "tfc:food/beet",
-    "tfc:food/cabbage",
-    "tfc:food/carrot",
-    "tfc:food/garlic",
-    "tfc:food/green_bean",
-    "tfc:food/green_bell_pepper",
-    "tfc:food/onion",
-    "tfc:food/potato",
-    "tfc:food/red_bell_pepper",
-    "tfc:food/dried_seaweed",
-    "tfc:food/dried_kelp",
-    "tfc:food/cattail_root",
-    "tfc:food/soybean",
-    "tfc:food/squash",
-    "tfc:food/tomato",
-    "tfc:food/yellow_bell_pepper",
-    "tfc:food/cheese",
-    "tfc:food/cooked_egg",
-    "tfc:food/beef",
-    "tfc:food/pork",
-    "tfc:food/chicken",
-    "tfc:food/mutton",
-    "tfc:food/bluegill",
-    "tfc:food/cod",
-    "tfc:food/salmon",
-    "tfc:food/tropical_fish",
-    "tfc:food/bear",
-    "tfc:food/horse_meat",
-    "tfc:food/pheasant",
-    "tfc:food/venison",
-    "tfc:food/wolf",
-    "tfc:food/rabbit",
-    "tfc:food/hyena",
-    "tfc:food/duck",
-    "tfc:food/chevon",
-    "tfc:food/gran_feline",
-    "tfc:food/camelidae",
-    "tfc:food/cooked_beef",
-    "tfc:food/cooked_pork",
-    "tfc:food/cooked_chicken",
-    "tfc:food/cooked_mutton",
-    "tfc:food/cooked_cod",
-    "tfc:food/cooked_tropical_fish",
-    "tfc:food/cooked_salmon",
-    "tfc:food/cooked_bluegill",
-    "tfc:food/cooked_bear",
-    "tfc:food/cooked_horse_meat",
-    "tfc:food/cooked_pheasant",
-    "tfc:food/cooked_venison",
-    "tfc:food/cooked_wolf",
-    "tfc:food/cooked_rabbit",
-    "tfc:food/cooked_hyena",
-    "tfc:food/cooked_duck",
-    "tfc:food/cooked_chevon",
-    "tfc:food/cooked_gran_feline",
-    "tfc:food/cooked_camelidae"
+    "#tfc:foods"
   ]
 }
